backend/query/requirements.txt
functions-framework==3.*
datasets==2.8.0
pinecone-client==2.1.0
openai==0.26.0